js updates to make the tile in offcanvas title
